@@ -1,7 +1,7 @@
 {
   "name": "contentstack-nextjs-starter-app",
   "description": "A starter app for Contentstack and Nextjs",
-  "version": "2.0.9",
+  "version": "2.1.0",
   "private": true,
   "author": "Contentstack",
   "scripts": {
@@ -11,26 +11,28 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@contentstack/live-preview-utils": "^1.4.0",
+    "@contentstack/live-preview-utils": "^3.2.2",
     "@contentstack/utils": "^1.3.1",
     "@textea/json-viewer": "^3.3.0",
-    "contentstack": "^3.19.0",
+    "contentstack": "^3.25.3",
     "html-react-parser": "^3.0.16",
     "moment": "^2.30.1",
-    "next": "^14.1.0",
+    "next": "^15.3.2",
     "next-pwa": "^5.6.0",
     "nprogress": "^0.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-loading-skeleton": "^3.3.1"
+    "react-loading-skeleton": "^3.5.0",
+    "sanitize-html": "^2.17.0"
   },
   "devDependencies": {
     "@types/node": "^20.11.5",
     "@types/nprogress": "^0.2.3",
     "@types/react": "^18.2.57",
     "@types/react-dom": "^18.2.19",
+    "@types/sanitize-html": "^2.16.0",
     "eslint": "^8.56.0",
-    "eslint-config-next": "^14.0.4",
+    "eslint-config-next": "^15.3.2",
     "typescript": "^5.3.3"
   }
 }
