@@ -1,89 +1,45 @@
-import { Config, Region, LivePreview, Stack } from "contentstack";
-import getConfig from "next/config";
-const { publicRuntimeConfig } = getConfig();
-const envConfig = process.env.CONTENTSTACK_API_KEY
-  ? process.env
-  : publicRuntimeConfig;
-const {
-  CONTENTSTACK_API_KEY,
-  CONTENTSTACK_DELIVERY_TOKEN,
-  CONTENTSTACK_ENVIRONMENT,
-  CONTENTSTACK_BRANCH,
-  CONTENTSTACK_REGION,
-  CONTENTSTACK_PREVIEW_TOKEN,
-  CONTENTSTACK_PREVIEW_HOST,
-  CONTENTSTACK_APP_HOST,
-  CONTENTSTACK_LIVE_PREVIEW,
-} = envConfig;
 
-// basic env validation
-export const isBasicConfigValid = () => {
-  return (
-    !!CONTENTSTACK_API_KEY &&
-    !!CONTENTSTACK_DELIVERY_TOKEN &&
-    !!CONTENTSTACK_ENVIRONMENT
-  );
-};
-// Live preview config validation
-export const isLpConfigValid = () => {
-  return (
-    !!CONTENTSTACK_LIVE_PREVIEW &&
-    !!CONTENTSTACK_PREVIEW_TOKEN &&
-    !!CONTENTSTACK_PREVIEW_HOST &&
-    !!CONTENTSTACK_APP_HOST
-  );
-};
-// set region
+import { Config, Region, Stack } from 'contentstack';
+import {publicConfig} from './config';
+
 const setRegion = (): Region => {
-  let region = "US" as keyof typeof Region;
-  if (!!CONTENTSTACK_REGION && CONTENTSTACK_REGION !== "us") {
-    region = CONTENTSTACK_REGION.toLocaleUpperCase().replace(
-      "-",
-      "_"
+  let region = 'US' as keyof typeof Region;
+  if (!!publicConfig.region && publicConfig.region !== 'us') {
+    region = publicConfig.region.toLocaleUpperCase().replace(
+      '-',
+      '_'
     ) as keyof typeof Region;
   }
   return Region[region];
 };
-// set LivePreview config
-const setLivePreviewConfig = (): LivePreview => {
-  if (!isLpConfigValid())
-    throw new Error("Your LP config is set to true. Please make you have set all required LP config in .env");
-  return {
-    preview_token: CONTENTSTACK_PREVIEW_TOKEN as string,
-    enable: CONTENTSTACK_LIVE_PREVIEW === "true",
-    host: CONTENTSTACK_PREVIEW_HOST as string,
-  } as LivePreview;
-};
-// contentstack sdk initialization
-export const initializeContentStackSdk = (): Stack => {
-  if (!isBasicConfigValid())
-    throw new Error("Please set you .env file before running starter app");
+
+export const initializeContentStackSdk = (config:any): Stack => {
   const stackConfig: Config = {
-    api_key: CONTENTSTACK_API_KEY as string,
-    delivery_token: CONTENTSTACK_DELIVERY_TOKEN as string,
-    environment: CONTENTSTACK_ENVIRONMENT as string,
+    api_key: config.apiKey as string,
+    delivery_token: config.deliveryToken as string,
+    environment: config.environment as string,
     region: setRegion(),
-    branch: CONTENTSTACK_BRANCH || "main",
+    branch: config.branch || 'main',
   };
-  if (CONTENTSTACK_LIVE_PREVIEW === "true") {
-    stackConfig.live_preview = setLivePreviewConfig();
-  }
+
+  stackConfig.live_preview = {
+    preview_token: config.previewToken as string,
+    enable: config.livePreview === 'true',
+    host: config.previewHost as string,
+  };
   return Stack(stackConfig);
 };
-// api host url
 export const customHostUrl = (baseUrl: string): string => {
-  return baseUrl.replace("api", "cdn");
+  return baseUrl.replace('api', 'cdn');
 };
-// generate prod api urls
 export const generateUrlBasedOnRegion = (): string[] => {
   return Object.keys(Region).map((region) => {
-    if (region === "US") {
+    if (region === 'US') {
       return `cdn.contentstack.io`;
     }
     return `${region}-cdn.contentstack.com`;
   });
 };
-// prod url validation for custom host
-export const isValidCustomHostUrl = (url=''): boolean => {
+export const isValidCustomHostUrl = (url = ''): boolean => {
   return url ? !generateUrlBasedOnRegion().includes(url) : false;
 };
