@@ -1,27 +1,32 @@
-const withPWA = require("next-pwa")({
-  dest: "public",
+const { env } = require('process');
+
+/** @type {import('next').NextConfig} */
+const withPWA = require('next-pwa')({
+  dest: 'public',
+  disable: process.env.NODE_ENV === 'development',
 });
 
-const config = {
-  publicRuntimeConfig: {
-    // Will be available on both server and client
-    CONTENTSTACK_API_KEY: process.env.CONTENTSTACK_API_KEY,
-    CONTENTSTACK_DELIVERY_TOKEN: process.env.CONTENTSTACK_DELIVERY_TOKEN,
-    CONTENTSTACK_BRANCH: process.env.CONTENTSTACK_BRANCH || "main",
-    CONTENTSTACK_REGION: process.env.CONTENTSTACK_REGION || "us",
-    CONTENTSTACK_ENVIRONMENT: process.env.CONTENTSTACK_ENVIRONMENT,
-    CONTENTSTACK_PREVIEW_TOKEN: process.env.CONTENTSTACK_PREVIEW_TOKEN,
-    CONTENTSTACK_PREVIEW_HOST:
-      process.env.CONTENTSTACK_PREVIEW_HOST || "rest-preview.contentstack.com",
-    CONTENTSTACK_API_HOST:
-      process.env.CONTENTSTACK_API_HOST || "api.contentstack.io",
-    CONTENTSTACK_APP_HOST:
-      process.env.CONTENTSTACK_APP_HOST || "app.contentstack.com",
-    CONTENTSTACK_LIVE_PREVIEW: process.env.CONTENTSTACK_LIVE_PREVIEW || "true",
-    CONTENTSTACK_LIVE_EDIT_TAGS:
-      process.env.CONTENTSTACK_LIVE_EDIT_TAGS || "false",
+const nextConfig = {
+  reactStrictMode: true,
+  swcMinify: true,
+  output: 'standalone',
+  images: {
+    formats: ['image/avif', 'image/webp'],
   },
-  experimental: { largePageDataBytes: 128 * 100000 },
+  env: {
+    apiKey: process.env.CONTENTSTACK_API_KEY,
+    deliveryToken: process.env.CONTENTSTACK_DELIVERY_TOKEN,
+    environment: process.env.CONTENTSTACK_ENVIRONMENT,
+    appHost: process.env.CONTENTSTACK_APP_HOST,
+    apiHost: process.env.CONTENTSTACK_API_HOST,
+    previewHost: process.env.CONTENTSTACK_PREVIEW_HOST,
+    region: process.env.CONTENTSTACK_REGION,
+    livePreview: process.env.CONTENTSTACK_LIVE_PREVIEW,
+    branch: process.env.CONTENTSTACK_BRANCH,
+    previewToken: process.env.CONTENTSTACK_PREVIEW_TOKEN,
+    liveEdit: process.env.CONTENTSTACK_LIVE_EDIT_TAGS,
+  },
 };
+
 module.exports =
-  process.env.NODE_ENV === "development" ? config : withPWA(config);
+  process.env.NODE_ENV === 'development' ? nextConfig : withPWA(nextConfig);
